Move lru-disk-cache to an internal module
See discussion in #916
