--- lru-disk-cache/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "lru-disk-cache"
-version = "0.4.1"
-authors = ["Ted Mielczarek <ted@mielczarek.org>"]
-license = "Apache-2.0"
-description = "A LRU cache for files on disk."
-repository = "https://github.com/mozilla/sccache/"
-edition = "2018"
-
-[dependencies]
-filetime = "0.2"
-linked-hash-map = "0.5"
-log = "0.4"
-# This is currently vendored to allow publishing on crates.io.
-#lru-cache = { git = "https://github.com/luser/lru-cache", branch = "non-mut-get" }
-walkdir = "2.3.1"
-
-[dev-dependencies]
-tempfile = "3"
