+++ src/bin/sccache-dist/build.rs
@@ -15,11 +15,11 @@
 use anyhow::{anyhow, bail, Context, Error, Result};
 use flate2::read::GzDecoder;
 use libmount::Overlay;
-use lru_disk_cache::Error as LruError;
 use sccache::dist::{
     BuildResult, BuilderIncoming, CompileCommand, InputsReader, OutputData, ProcessOutput, TcCache,
     Toolchain,
 };
+use sccache::lru_disk_cache::Error as LruError;
 use std::collections::{hash_map, HashMap};
 use std::fs;
 use std::io;
