+++ src/compiler/rust.rs
@@ -22,6 +22,8 @@ use crate::compiler::{DistPackagers, OutputsRewriter};
 use crate::dist;
 #[cfg(feature = "dist-client")]
 use crate::dist::pkg;
+#[cfg(feature = "dist-client")]
+use crate::lru_disk_cache::{LruCache, Meter};
 use crate::mock_command::{CommandCreatorSync, RunCommand};
 use crate::util::{fmt_duration_as_secs, hash_all, run_input_output, Digest};
 use crate::util::{ref_env, HashToDigest, OsStrExt, SpawnExt};
@@ -29,8 +31,6 @@ use filetime::FileTime;
 use futures::Future;
 use futures_03::executor::ThreadPool;
 use log::Level::Trace;
-#[cfg(feature = "dist-client")]
-use lru_disk_cache::{LruCache, Meter};
 #[cfg(feature = "dist-client")]
 #[cfg(feature = "dist-client")]
 use std::borrow::Borrow;
