+++ src/lib.rs
@@ -51,6 +51,7 @@ mod compiler;
 pub mod config;
 pub mod dist;
 mod jobserver;
+pub mod lru_disk_cache;
 mod mock_command;
 mod protocol;
 pub mod server;
